+++ src/pint/models/astrometry.py
@@ -63,13 +63,19 @@ class Astrometry(DelayComponent):
         # TODO: would it be better for this to return a 6-vector (pos, vel)?
         return self.coords_as_ICRS(epoch=epoch).cartesian.xyz.transpose()
 
-    def ssb_to_psb_xyz_ECL(self, epoch=None):
+    def ssb_to_psb_xyz_ECL(self, epoch=None, ecl=None):
         """Returns unit vector(s) from SSB to pulsar system barycenter under Ecliptic coordinates.
 
         If epochs (MJD) are given, proper motion is included in the calculation.
+
+        Parameters
+        ----------
+        epoch : float, optional
+        ecl : str, optional
+            Obliquity (IERS2010 by default)
         """
         # TODO: would it be better for this to return a 6-vector (pos, vel)?
-        return self.coords_as_ECL(epoch=epoch).cartesian.xyz.transpose()
+        return self.coords_as_ECL(epoch=epoch, ecl=ecl).cartesian.xyz.transpose()
 
     def sun_angle(self, toas, heliocenter=True, also_distance=False):
         """Compute the pulsar-observatory-Sun angle.
@@ -1011,8 +1017,8 @@ class AstrometryEcliptic(Astrometry):
             frame=PulsarEcliptic,
         )
         c_ICRS = c.transform_to(coords.ICRS)
-        m_eq.RAJ.uncertainty = c_ICRS.pm_ra_cosdec * dt / np.cos(c_ICRS.dec)
-        m_eq.DECJ.uncertainty = c_ICRS.pm_dec * dt
+        m_eq.RAJ.uncertainty = np.abs(c_ICRS.pm_ra_cosdec * dt / np.cos(c_ICRS.dec))
+        m_eq.DECJ.uncertainty = np.abs(c_ICRS.pm_dec * dt)
         # use fake proper motions to convert uncertainties on proper motion
         # assume that PMELONG uncertainty includes cos(DEC)
         c = coords.SkyCoord(
@@ -1025,8 +1031,8 @@ class AstrometryEcliptic(Astrometry):
             frame=PulsarEcliptic,
         )
         c_ICRS = c.transform_to(coords.ICRS)
-        m_eq.PMRA.uncertainty = c_ICRS.pm_ra_cosdec
-        m_eq.PMDEC.uncertainty = c_ICRS.pm_dec
+        m_eq.PMRA.uncertainty = np.abs(c_ICRS.pm_ra_cosdec)
+        m_eq.PMDEC.uncertainty = np.abs(c_ICRS.pm_dec)
         # freeze comparable parameters
         m_eq.RAJ.frozen = self.ELONG.frozen
         m_eq.DECJ.frozen = self.ELAT.frozen