+++ src/pint/scripts/pintk.py
@@ -9,6 +9,7 @@ import tkinter as tk
 import tkinter.filedialog as tkFileDialog
 import tkinter.messagebox as tkMessageBox
 from tkinter import ttk
+import matplotlib as mpl
 
 import pint.logging
 from loguru import logger as log
@@ -208,7 +209,7 @@ class PINTk:
     def about(self):
         tkMessageBox.showinfo(
             title="About PINTk",
-            message=f"A Tkinter based graphical interface to PINT (version={pint.__version__})",
+            message=f"A Tkinter based graphical interface to PINT (version={pint.__version__}), using matplotlib (version={mpl.__version__}) and the {mpl.get_backend()} backend",
         )
 
 
@@ -247,9 +248,9 @@ def main(argv=None):
     parser.add_argument(
         "-v",
         "--version",
-        default=False,
-        action="store_true",
+        action="version",
         help="Print version info and  exit.",
+        version=f"This is PINT version {pint.__version__}, using matplotlib (version={mpl.__version__})",
     )
     parser.add_argument(
         "--log-level",
@@ -260,10 +261,6 @@ def main(argv=None):
         dest="loglevel",
     )
     args = parser.parse_args(argv)
-
-    if args.version:
-        print(f"This is PINT version {pint.__version__}")
-        sys.exit(0)
 
     if args.loglevel != "WARNING":
         log.remove()